requirements.txt
uvicorn==0.24.0.post1
starlette==0.32.0.post1
python-telegram-bot==20.6
requests
dotenv
httpx